--- great_expectations/metrics/column_map/__init__.py
@@ -1,1 +1,0 @@
-from .between import Between
